# specs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gaia Application Framework API",
    "version": "0.1.0",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "paths": {
...
    "/actuator/components": {
      "get": {
        "tags": [
          "actuator"
        ],
        "summary": "Components",
        "operationId": "components_actuator_components_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Components Actuator Components Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/actuator/config": {
      "get": {
        "tags": [
          "actuator"
        ],
        "summary": "Config",
        "operationId": "config_actuator_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActuatorConfig"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/actuator/conditions": {
      "get": {
        "tags": [
          "actuator"
        ],
        "summary": "Conditions",
        "operationId": "conditions_actuator_conditions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActuatorCondition"
                  },
                  "type": "array",
                  "title": "Response Conditions Actuator Conditions Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/actuator/runtime": {
      "get": {
        "tags": [
          "actuator"
        ],
        "summary": "Runtime Summary",
        "operationId": "runtime_summary_actuator_runtime_get",
        "parameters": [
          {
            "name": "window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Window Hours"
            }
          },
          {
            "name": "stale_after_seconds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 86400,
              "minimum": 30,
              "default": 300,
              "title": "Stale After Seconds"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActuatorCondition": {
        "properties": {
          "starter_id": {
            "type": "string",
            "title": "Starter Id"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "starter_id",
          "matched",
          "reasons"
        ],
        "title": "ActuatorCondition"
      },
      "ActuatorConfig": {
        "properties": {
          "config_hash": {
            "type": "string",
            "title": "Config Hash"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "origins": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Origins"
          }
        },
        "type": "object",
        "required": [
          "config_hash",
          "config",
          "origins"
        ],
        "title": "ActuatorConfig"
      },
...
      "DatabaseContention": {
        "properties": {
          "backend": {
            "type": "string",
            "title": "Backend"
          },
          "pool_class": {
            "type": "string",
            "title": "Pool Class"
          },
          "pool_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pool Size"
          },
          "checked_out": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Out"
          },
          "overflow": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overflow"
          },
          "waiting_connections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiting Connections"
          },
          "lock_waiting_connections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lock Waiting Connections"
          }
        },
        "type": "object",
        "required": [
          "backend",
          "pool_class"
        ],
        "title": "DatabaseContention"
      },
...
      "DurationSummary": {
        "properties": {
          "average_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Ms"
          },
          "p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Ms"
          },
          "p95_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ms"
          }
        },
        "type": "object",
        "title": "DurationSummary"
      },
      "ErrorCategory": {
        "type": "string",
        "enum": [
          "request",
          "authentication",
          "authorization",
          "configuration",
          "conflict",
          "external_dependency",
          "policy",
          "runtime",
          "internal",
          "unknown"
        ],
        "title": "ErrorCategory"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "INVALID_REQUEST",
          "SCENARIO_NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "POLICY_DENIED",
          "ENVIRONMENT_MODE_MISMATCH",
          "TOOL_NOT_REGISTERED",
          "TOOL_NOT_ALLOWED",
          "TOOL_ENVIRONMENT_MISMATCH",
          "TOOL_DEFINITION_MISMATCH",
          "TOOL_ROLE_REQUIRED",
          "TOOL_ADAPTER_ERROR",
          "WRITE_DISABLED",
          "CONTEXT_INSUFFICIENT",
          "MODEL_CAPABILITY_MISSING",
          "MODEL_UNAVAILABLE",
          "TOOL_TIMEOUT",
          "SIDE_EFFECT_UNKNOWN",
          "HUMAN_GATE_REJECTED",
          "HUMAN_GATE_EXPIRED",
          "BUDGET_EXCEEDED",
          "RUN_NOT_RESUMABLE",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCategory"
              },
              {
                "type": "string"
              }
            ],
            "title": "Category",
            "default": "unknown"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable",
            "default": false
          },
          "operator_action": {
            "type": "string",
            "title": "Operator Action",
            "default": "Inspect the diagnostic trace before retrying."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutboxSummary": {
        "properties": {
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          },
          "pending": {
            "type": "integer",
            "title": "Pending",
            "default": 0
          },
          "retrying": {
            "type": "integer",
            "title": "Retrying",
            "default": 0
          },
          "dead_letter": {
            "type": "integer",
            "title": "Dead Letter",
            "default": 0
          }
        },
        "type": "object",
        "title": "OutboxSummary"
      },
...
      "RuntimeIssue": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "bottleneck": {
            "type": "string",
            "title": "Bottleneck"
          },
          "age_seconds": {
            "type": "integer",
            "title": "Age Seconds"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "scenario_id",
          "status",
          "bottleneck",
          "age_seconds",
          "updated_at"
        ],
        "title": "RuntimeIssue"
      },
      "RuntimeSummary": {
        "properties": {
          "window_hours": {
            "type": "integer",
            "title": "Window Hours"
          },
          "stale_after_seconds": {
            "type": "integer",
            "title": "Stale After Seconds"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "total_runs": {
            "type": "integer",
            "title": "Total Runs"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Status Counts"
          },
          "success_rate": {
            "type": "number",
            "title": "Success Rate"
          },
          "failure_rate": {
            "type": "number",
            "title": "Failure Rate"
          },
          "blocked_rate": {
            "type": "number",
            "title": "Blocked Rate"
          },
          "active_runs": {
            "type": "integer",
            "title": "Active Runs"
          },
          "stale_runs": {
            "type": "integer",
            "title": "Stale Runs"
          },
          "pending_human_gates": {
            "type": "integer",
            "title": "Pending Human Gates"
          },
          "oldest_pending_gate_age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Pending Gate Age Seconds"
          },
          "run_duration": {
            "$ref": "#/components/schemas/DurationSummary"
          },
          "human_gate_wait": {
            "$ref": "#/components/schemas/DurationSummary"
          },
          "error_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Error Counts"
          },
          "database": {
            "$ref": "#/components/schemas/DatabaseContention"
          },
          "outbox": {
            "$ref": "#/components/schemas/OutboxSummary"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/RuntimeIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "window_hours",
          "stale_after_seconds",
          "generated_at",
          "total_runs",
          "status_counts",
          "success_rate",
          "failure_rate",
          "blocked_rate",
          "active_runs",
          "stale_runs",
          "pending_human_gates",
          "oldest_pending_gate_age_seconds",
          "run_duration",
          "human_gate_wait",
          "error_counts",
          "database",
          "outbox",
          "issues"
        ],
        "title": "RuntimeSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Gaia-Api-Key"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Gaia application on the current host"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}